+++ packages/terra-framework-docs/src/terra-dev-site/doc/folder-tree/Examples.2/BasicFolderTree.jsx
@@ -8,12 +8,48 @@ import styles from './BasicFolderTree.module.scss';
 
 const cx = classNames.bind(styles);
 
+const data = {
+  title: 'Documents',
+  items: [
+    { id: 'item-1-level-1', label: 'Information', icon: <IconDocuments /> },
+    { id: 'item-2-level-1', label: 'Projects - 2023', icon: <IconDocuments /> },
+    {
+      id: 'item-3-level-1',
+      label: 'Projects - 2023',
+      items: [
+        { id: 'item-1-level-2-1', label: 'Project Data - October', icon: <IconDocuments /> },
+        { id: 'item-2-level-2-1', label: 'Project Data - November', icon: <IconDocuments /> },
+        { id: 'item-3-level-2-1', label: 'Project Data - December', icon: <IconDocuments /> },
+        {
+          id: 'item-4-level-2-1',
+          label: 'Tests',
+          items: [
+            { id: 'item-1-level-3-1', label: 'Very Very Very Very Very Very Very Long Name Test', icon: <IconDocuments /> },
+            { id: 'item-1-level-2-1', label: 'Even Looooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooonger Name Test', icon: <IconDocuments /> },
+          ],
+        },
+      ],
+    },
+    { id: 'item-4-level-1', label: 'Projects - 2024', icon: <IconDocuments /> },
+    {
+      id: 'item-5-level-1',
+      label: 'Projects - 2024',
+      items: [
+        { id: 'item-1-level-2-2', label: 'Project Data - January', icon: <IconDocuments /> },
+      ],
+    },
+  ],
+};
+
+const folderExpandedState = {
+  'item-3-level-1': false,
+  'item-4-level-2-1': false,
+  'item-5-level-1': false,
+};
+
 const BasicFolderTree = () => {
-  const [selectedKey, setSelectedKey] = React.useState('info');
-  const [expandedItems, setExpandedItems] = React.useState({
-    projects: false,
-    tests: false,
-  });
+  const [selectedKey, setSelectedKey] = React.useState('');
+  const [expandedItems, setExpandedItems] = React.useState(folderExpandedState);
 
   const handleExpandCollapseKeys = (key) => {
     const newExpandedItems = {
@@ -41,84 +77,37 @@ const BasicFolderTree = () => {
     setExpandedItems(newExpandedItems);
   };
 
+  const buildFolderTreeItems = (items) => {
+    if (items) {
+      const elements = [];
+      items.forEach(item => {
+        elements.push(
+          <FolderTree.Item
+            key={item.id}
+            label={item.label}
+            icon={item.icon}
+            subfolderItems={buildFolderTreeItems(item.items)}
+            isExpanded={expandedItems[item.id]}
+            isSelected={selectedKey === item.id}
+            onClick={() => { setSelectedKey(item.id); }}
+            onToggle={() => { handleExpandCollapseKeys(item.id); }}
+          />,
+        );
+      });
+      return elements;
+    }
+
+    return null;
+  };
+
   return (
     <div className={cx('content-wrapper')}>
       <FolderTree
-        title="Documents"
-        key="documents"
+        title={data.title}
         onExpandAll={handleExpandAll}
         onCollapseAll={handleCollapseAll}
       >
-        <FolderTree.Item
-          label="info.txt"
-          icon={<IconDocuments />}
-          key="info"
-          isSelected={selectedKey === 'info'}
-          onClick={() => { setSelectedKey('info'); }}
-        />
-        <FolderTree.Item
-          label="projects_2023.txt"
-          icon={<IconDocuments />}
-          key="projects_2023"
-          isSelected={selectedKey === 'projects_2023'}
-          onClick={() => { setSelectedKey('projects_2023'); }}
-        />
-        <FolderTree.Item
-          label="Projects"
-          key="projects"
-          isSelected={selectedKey === 'projects'}
-          isExpanded={expandedItems.projects}
-          onClick={() => { setSelectedKey('projects'); }}
-          onToggle={() => { handleExpandCollapseKeys('projects'); }}
-          subfolderItems={[
-            (<FolderTree.Item
-              label="project_data1.txt"
-              icon={<IconDocuments />}
-              key="project_data1"
-              isSelected={selectedKey === 'project_data1'}
-              onClick={() => { setSelectedKey('project_data1'); }}
-            />),
-            (<FolderTree.Item
-              label="project_data2.txt"
-              icon={<IconDocuments />}
-              key="project_data2"
-              isSelected={selectedKey === 'project_data2'}
-              onClick={() => { setSelectedKey('project_data2'); }}
-            />),
-            (<FolderTree.Item
-              label="Tests"
-              key="tests"
-              isSelected={selectedKey === 'tests'}
-              isExpanded={expandedItems.tests}
-              onClick={() => { setSelectedKey('tests'); }}
-              onToggle={() => { handleExpandCollapseKeys('tests'); }}
-              icon={<IconDocuments />}
-              subfolderItems={[
-                (<FolderTree.Item
-                  label="very_very_very_very_very_very_very_long_name_test.txt"
-                  icon={<IconDocuments />}
-                  key="long-name-test"
-                  isSelected={selectedKey === 'long-name-test'}
-                  onClick={() => { setSelectedKey('long-name-test'); }}
-                />),
-                (<FolderTree.Item
-                  label="test.txt"
-                  icon={<IconDocuments />}
-                  key="test"
-                  isSelected={selectedKey === 'test'}
-                  onClick={() => { setSelectedKey('test'); }}
-                />),
-                (<FolderTree.Item
-                  label="even_longer_long_long_long_long_long_long_long_long_long_long_long_long_name_test.txt"
-                  icon={<IconDocuments />}
-                  key="longer-name-test"
-                  isSelected={selectedKey === 'longer-name-test'}
-                  onClick={() => { setSelectedKey('longer-name-test'); }}
-                />),
-              ]}
-            />),
-          ]}
-        />
+        {buildFolderTreeItems(data.items)}
       </FolderTree>
     </div>
   );