+++ packages/terra-framework-docs/src/terra-dev-site/doc/pills/example/VisibleLabelExample.jsx
@@ -29,12 +29,15 @@ const VisibleLabelExample = () => {
     },
   ];
   const [pills, setPills] = useState(pillsData);
-  const isResetButtonDisabled = (pills.length === pillsData.length);
 
   const handleOnRemove = (id, metaData) => {
     const pillsArray = pills;
     pillsArray.splice(metaData.index, 1);
     setPills([...pillsArray]);
+  };
+
+  const handleResetClick = () => {
+    setPills(pillsData);
   };
 
   const patientEducationSearchTermsLabel = 'Patient Education Search Terms:';
@@ -66,10 +69,8 @@ const VisibleLabelExample = () => {
         </FilterPills>
       </div>
       <DocsButton
-        aria-disabled={isResetButtonDisabled}
-        className={cx(['terra-docs-example-reset-button', { 'is-disabled': isResetButtonDisabled }, theme.className])}
-        disabled={isResetButtonDisabled}
-        onClick={() => setPills(pillsData)}
+        className={cx(['terra-docs-example-reset-button', theme.className])}
+        onClick={handleResetClick}
       >
         Reset Example
       </DocsButton>