+++ packages/terra-pills/src/FilterPills.jsx
@@ -108,7 +108,7 @@ const FilterPills = (props) => {
 
   // Modifies the tabindex of the pill/rollup pill node
   const setTabIndex = (val) => {
-    const currentNode = currentPill.current ? filterPillsRef.current.querySelector(`#${currentPill.current}`) : null;
+    const currentNode = currentPill.current ? filterPillsRef.current.querySelector(`[id='${currentPill.current}']`) : null;
     if (currentNode) {
       currentNode.setAttribute('tabIndex', val);
     }
@@ -116,7 +116,7 @@ const FilterPills = (props) => {
 
   // Sets focus to the current pill/rollup pill with tabindex 0
   const focusCurrentNode = () => {
-    const currentNode = currentPill.current ? filterPillsRef.current.querySelector(`#${currentPill.current}`) : null;
+    const currentNode = currentPill.current ? filterPillsRef.current.querySelector(`[id='${currentPill.current}']`) : null;
     if (currentNode) {
       currentNode.focus();
     }
@@ -404,31 +404,37 @@ const FilterPills = (props) => {
     'aria-labelledby': ariaLabelledBy,
     'aria-describedby': pillGroupAriaDescribedBy,
     role: 'list',
-    tabIndex: containerTabindex,
   } : { role: 'group' };
 
+  if (!React.Children.count(children)) {
+    pillGroupInteractionHint = `${pillGroupInteractionHint}, ${ariaLabelHint}`;
+  }
+
   return (
-    <div
-      {...customProps}
-      {...filterPillsProps}
-      {...ariaAttrs}
-      className={pillListClassNames}
-      ref={filterPillsRef}
-    >
-      <VisuallyHiddenText
-        aria-live="polite"
-        id={pillGroupInteractionHintID}
-        text={pillGroupInteractionHint}
-      />
-      {children ? renderChildren(children) : []}
-      {(isCollapsible && rollUpCount > 0) && (
+    <>
+      <div
+        {...customProps}
+        {...filterPillsProps}
+        {...ariaAttrs}
+        tabIndex={containerTabindex}
+        className={pillListClassNames}
+        ref={filterPillsRef}
+      >
+        {children ? renderChildren(children) : []}
+        {(isCollapsible && rollUpCount > 0) && (
         <RollUpPill
           isCollapsed={isCollapsed}
           onSelectRollUp={handleOnSelectRollUp}
           rollupCount={rollUpCount}
         />
-      )}
-    </div>
+        )}
+      </div>
+      <VisuallyHiddenText
+        aria-live="polite"
+        id={pillGroupInteractionHintID}
+        text={pillGroupInteractionHint}
+      />
+    </>
   );
 };