+++ packages/terra-collapsible-menu-view/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 6.87.1 - (January 10, 2024)
+
+* Fixed
+  * Fixed flicker issue when `collapsible-menu-view` is wrapped with `terra-clinical-header`.
+
 ## 6.87.0 - (December 18, 2023)
 
 * Added