+++ packages/terra-compact-interactive-list/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 1.0.1 - (January 10, 2024)
+
+* Fixed
+  * Cell does not call `onCellSelect` method if corresponding column `isSelectable` prop set to `false`.
+
 ## 1.0.0 - (January 2, 2024)
 
 * Changed